+++ bitalarm-impl/src/main/java/com/dwarfeng/bitalarm/impl/service/operation/AlarmSettingCrudOperation.java
@@ -1,6 +1,5 @@
 package com.dwarfeng.bitalarm.impl.service.operation;
 
-import com.dwarfeng.bitalarm.stack.bean.entity.AlarmHistory;
 import com.dwarfeng.bitalarm.stack.bean.entity.AlarmSetting;
 import com.dwarfeng.bitalarm.stack.cache.AlarmSettingCache;
 import com.dwarfeng.bitalarm.stack.cache.EnabledAlarmSettingCache;
@@ -8,7 +7,6 @@ import com.dwarfeng.bitalarm.stack.dao.AlarmHistoryDao;
 import com.dwarfeng.bitalarm.stack.dao.AlarmInfoDao;
 import com.dwarfeng.bitalarm.stack.dao.AlarmSettingDao;
 import com.dwarfeng.bitalarm.stack.dao.CurrentAlarmDao;
-import com.dwarfeng.bitalarm.stack.service.AlarmHistoryMaintainService;
 import com.dwarfeng.subgrade.sdk.exception.ServiceExceptionCodes;
 import com.dwarfeng.subgrade.sdk.service.custom.operation.CrudOperation;
 import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
@@ -16,9 +14,6 @@ import com.dwarfeng.subgrade.stack.exception.ServiceException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Component
 public class AlarmSettingCrudOperation implements CrudOperation<LongIdKey, AlarmSetting> {
@@ -96,17 +91,6 @@ public class AlarmSettingCrudOperation implements CrudOperation<LongIdKey, Alarm
             if (alarmInfoDao.exists(key)) {
                 alarmInfoDao.delete(key);
             }
-        }
-
-        //删除与点位相关的过滤器触发器。
-        {
-            //查找点位拥有的过滤器与触发器。
-            List<LongIdKey> alarmHistoryKeys = alarmHistoryDao
-                    .lookup(AlarmHistoryMaintainService.CHILD_FOR_ALARM_SETTING, new Object[]{key})
-                    .stream().map(AlarmHistory::getKey).collect(Collectors.toList());
-
-            //删除点位拥有的过滤器与触发器。
-            alarmHistoryDao.batchDelete(alarmHistoryKeys);
         }
 
         alarmSettingDao.delete(key);